+++ docs/website/docs/getting-started-snippets.py
@@ -117,7 +117,7 @@ def db_snippet() -> None:
         pipeline = dlt.pipeline(
             pipeline_name='from_database',
             destination='duckdb',
-            dataset_name='genome_da ta',
+            dataset_name='genome_data',
         )
 
         # here we convert the rows into dictionaries on the fly with a map function
